pyproject.toml
[project]
name = "vl-computer-use"
version = "0.1.0"
description = "Vision Language Model Computer Use with RL"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "torch",
    "torchvision",
    "transformers>=4.50.0",
    "bitsandbytes>=0.49.0",
    "accelerate>=1.0.0",
    "peft>=0.18.0",
    "huggingface-hub>=0.36.0",
    "sentencepiece",
    "protobuf",
    "qwen-vl-utils",
    "aiohttp>=3.13.3",
    "websockets>=16.0",
    "playwright>=1.58.0",
    "requests",
    "pillow",
    "wandb",
    "vllm>=0.8.0",
    "openai",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
